+++ Quake/client.h
@@ -68,7 +68,7 @@ typedef struct
 
 #define	SIGNONS		4			// signon messages to receive before connected
 
-#define	MAX_DLIGHTS		32
+#define	MAX_DLIGHTS		64
 typedef struct
 {
 	vec3_t	origin;
@@ -83,7 +83,7 @@ typedef struct
 } dlight_t;
 
 
-#define	MAX_BEAMS	24
+#define	MAX_BEAMS	64
 typedef struct
 {
 	int		entity;
@@ -92,7 +92,7 @@ typedef struct
 	vec3_t	start, end;
 } beam_t;
 
-#define	MAX_EFRAGS		640
+#define	MAX_EFRAGS		4096
 
 #define	MAX_MAPSTRING	2048
 #define	MAX_DEMOS		8
@@ -269,8 +269,8 @@ extern	cvar_t	m_forward;
 extern	cvar_t	m_side;
 
 
-#define	MAX_TEMP_ENTITIES	64			// lightning bolts, etc
-#define	MAX_STATIC_ENTITIES	128			// torches, etc
+#define	MAX_TEMP_ENTITIES	256			// lightning bolts, etc
+#define	MAX_STATIC_ENTITIES	1024		// torches, etc
 
 extern	client_state_t	cl;
 
@@ -303,7 +303,7 @@ void CL_Disconnect (void);
 void CL_Disconnect_f (void);
 void CL_NextDemo (void);
 
-#define			MAX_VISEDICTS	256
+#define			MAX_VISEDICTS	2048
 extern	int				cl_numvisedicts;
 extern	entity_t		*cl_visedicts[MAX_VISEDICTS];