+++ Quake/r_shared.h
@@ -65,11 +65,11 @@ extern vec3_t vpn, base_vpn;
 extern vec3_t vright, base_vright;
 extern entity_t *currententity;
 
-#define NUMSTACKEDGES 2400
+#define NUMSTACKEDGES 65536
 #define MINEDGES NUMSTACKEDGES
-#define NUMSTACKSURFACES 800
+#define NUMSTACKSURFACES 16384
 #define MINSURFACES NUMSTACKSURFACES
-#define MAXSPANS 3000
+#define MAXSPANS 65536
 
 // !!! if this is changed, it must be changed in asm_draw.h too !!!
 typedef struct espan_s {